pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "claude-feishu-flow"
version = "0.1.0"
description = "ChatOps-style experiment management: Feishu + Claude AI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "anthropic>=0.25.0",
    "pydantic-settings>=2.2.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "python-dotenv>=1.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["claude_feishu_flow*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
